+++ packages/lsp-server/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.1](https://github.com/Microsoft/vscode-extension-samples/compare/v0.17.1-alpha.1...v0.17.1) (2020-11-26)
+
+**Note:** Version bump only for package @dendronhq/lsp-server
+
+## [0.17.1-alpha.0](https://github.com/Microsoft/vscode-extension-samples/compare/v0.17.0...v0.17.1-alpha.0) (2020-11-25)
+
+**Note:** Version bump only for package @dendronhq/lsp-server
+
 # [0.17.0](https://github.com/Microsoft/vscode-extension-samples/compare/v0.16.3-alpha.5...v0.17.0) (2020-11-22)
 
 **Note:** Version bump only for package @dendronhq/lsp-server